src/player: Rewrite internal time player standard
* src/formatter/FormatterMediator.cpp
* src/formatter/FormatterMediator.h
* src/formatter/NominalEventMonitor.cpp
* src/player/AVPlayer.cpp
* src/player/AVPlayer.h
* src/player/ChannelPlayer.cpp
* src/player/ChannelPlayer.h
* src/player/INCLPlayer.h
* src/player/IPlayer.h
* src/player/Player.cpp
* src/player/Player.h

+++ src/player/ChannelPlayer.h
@@ -40,7 +40,7 @@ public:
   map<string, IPlayer *> *getPlayerMap ();
   IPlayer *getPlayer (const string &objectId);
   void select (IPlayer *selObject);
-  double getMediaTime ();
+  guint32 getMediaTime ();
 
 private:
   void setSurfacesParent (SDLWindow* parent);
